changed guest sweep to only keep those guests less than two hours old

+++ lib/tasks/scheduler.rake
@@ -48,7 +48,9 @@ task :sweep_guest_users => :environment do
   puts "sweep of guests"
   guests = User.where(is_guest: true).where(name: "guest")
   guests.each do |g|
-    g.destroy
+    if g.updated_at < (Time.now - 2.hours)
+      g.destroy
+    end
   end
 end
